% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGvizTracks.R
\name{createGenemodels}
\alias{createGenemodels}
\title{Generate gene model GRanges object from GTF file}
\usage{
createGenemodels(gtfFile)
}
\arguments{
\item{gtfFile}{GTF file (in Ensembl format)}
}
\description{
Generate gene model GRanges object from GTF file
}
\author{
Charlotte Soneson
}
