% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGvizTracks.R
\name{plotGvizTracks}
\alias{plotGvizTracks}
\title{Plot coverage and gene model tracks}
\usage{
plotGvizTracks(showGene = NULL, geneModels = NULL, geneModels2 = NULL,
  gtfFile = NULL, bwFiles = NULL, bwCond = NULL, showChr = NULL,
  minCoord = NULL, maxCoord = NULL, pdfFilename = NULL, pdfWidth = 7,
  pdfHeight = 7, ...)
}
\arguments{
\item{showGene}{Gene name or symbol for the gene to show in the plot.}

\item{geneModels}{GRanges object with gene models, typically generated from a
gtf file (e.g. with the \code{createGenemodels} function).}

\item{geneModels2}{Second GRanges object with gene models, e.g. those that
were excluded from \code{geneModels}.}

\item{gtfFile}{GTF file from which \code{geneModels} can be generated if it
is not explicitly provided.}

\item{bwFiles}{Named vector with paths to bigWig files. These will be used to
generate coverage plots.}

\item{bwCond}{Named vector corresponding to \code{bwFiles}, giving the group
label for each bigWig file. These will be used to color the coverage plots
by group.}

\item{showChr}{Name of the chromosome to show in the plot, if \code{showGene}
is not specified.}

\item{minCoord, maxCoord}{Minimum and maximum genomic coordinate to show, if
\code{showGene} is not specified.}

\item{pdfFilename}{File name of pdf file to which the plot should be saved.
If set to \code{NULL}, the plot is written to the current graphics device.}

\item{pdfWidth, pdfHeight}{Width and height of the pdf file to which the plot
should be saved.}

\item{...}{Additional arguments (not currently used)}
}
\description{
Plot coverage and gene model tracks
}
\author{
Charlotte Soneson
}
