% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iResViewer.R
\name{iResViewer}
\alias{iResViewer}
\title{Interactive visualization of DE results}
\usage{
iResViewer(results = list(), dimred = list(), genemodels = NULL,
  bwFiles = NULL, bwCond = NULL, appTitle = "iResViewer")
}
\arguments{
\item{results}{List of result data frames}

\item{dimred}{List of dimension reduction results}

\item{genemodels}{Gene models}

\item{bwFiles}{Vector of bigWig files}

\item{bwCond}{Vector of conditions for bigWig files}

\item{appTitle}{App title}
}
\description{
Interactive visualization of DE results
}
